Split up banner tests and tweak abbreviations
